+++ packages/better-auth/src/plugins/index.ts
